restructured Polls module 2.0
