+++ src/components/base/BaseCooperationPoll.vue
@@ -66,11 +66,11 @@
 <script lang="ts">
 import { defineComponent, VueElement } from 'vue';
 import { mapGetters } from 'vuex';
-import { PollModelWithResults } from '@/store/polls/models/poll.model';
+import { PollModelWithResults } from '@/polls/models/poll.model';
 import Button from 'primevue/button';
 import Menu from 'primevue/menu';
 import { StoreModuleEnum } from '@/store/types';
-import { DeletePollPayloadInterface, PollsActionEnum, PollStatusEnum, PollStatusType } from '@/store/polls/types';
+import { DeletePollPayloadInterface, PollsActionEnum, PollStatusEnum, PollStatusType } from '@/polls/store/types';
 import { CooperationGettersEnum } from '@/cooperation/store/types';
 import ConfirmPopup from 'primevue/confirmpopup';
 import Dialog from 'primevue/dialog';
