+++ src/polls/CooperationPolls.vue
@@ -34,9 +34,9 @@ import { defineComponent } from 'vue';
 import Button from 'primevue/button';
 import BaseCooperationPoll from '@/components/base/BaseCooperationPoll.vue';
 import CreatePollForm from '@/components/CreatePollForm.vue';
-import { PollModel } from '@/store/polls/models/poll.model';
+import { PollModel } from '@/polls/models/poll.model';
 import { StoreModuleEnum } from '@/store/types';
-import { PollsActionEnum, PollsGettersEnum } from '@/store/polls/types';
+import { PollsActionEnum, PollsGettersEnum } from '@/polls/store/types';
 import Dialog from 'primevue/dialog';
 import { CooperationGettersEnum } from '@/cooperation/store/types';
