+++ src/polls/store/getters.ts
@@ -1,7 +1,7 @@
 import { GetterTree } from 'vuex';
 
 import { RootStateInterface } from '@/store/types';
-import { PollsStateInterface, Getters } from '@/store/polls/types';
+import { PollsStateInterface, Getters } from '@/polls/store/types';
 
 export const getters: GetterTree<PollsStateInterface, RootStateInterface> & Getters = {
   getPolls: (state) => {
