+++ src/polls/store/index.ts
@@ -1,9 +1,9 @@
 import { Module } from 'vuex';
 import { RootStateInterface } from '@/store/types';
-import { PollsStateInterface, PollAcceptanceCriteriaEnum } from '@/store/polls/types';
-import { getters } from '@/store/polls/getters';
-import { mutations } from '@/store/polls/mutations';
-import { actions } from '@/store/polls/actions';
+import { PollsStateInterface, PollAcceptanceCriteriaEnum } from '@/polls/store/types';
+import { getters } from '@/polls/store/getters';
+import { mutations } from '@/polls/store/mutations';
+import { actions } from '@/polls/store/actions';
 
 export const state: PollsStateInterface = {
   cooperationPolls: [],
