+++ src/router/index.ts
@@ -10,8 +10,8 @@ import MainPage from '@/views/MainPage.vue';
 import CooperationInfo from '@/cooperation/components/CooperationInfo.vue';
 import PageNotFound from '@/views/PageNotFound.vue';
 import PageErrorInvintation from '@/views/PageErrorInvintation.vue';
-import CooperationPolls from '@/views/CooperationPolls.vue';
-import PollInfo from '@/views/PollInfo.vue';
+import CooperationPolls from '@/polls/CooperationPolls.vue';
+import PollInfo from '@/polls/PollInfo.vue';
 import InvitationSection from '@/views/InvitationSection.vue';
 
 import ManageApartments from '@/apartment/ManageApartments.vue';
